Move node update on layer deletion to method and unhook it
